refresh notes

+++ lib/widgets/custom_note_item.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:notes_app/cubits/show_notes_cubit/show_notes_cubit.dart';
 import 'package:notes_app/models/note_model.dart';
 import 'package:notes_app/views/edit_note_view.dart';
 
@@ -40,8 +42,31 @@ class CustomNoteItem extends StatelessWidget {
                 ),
               ),
               trailing: IconButton(
-                onPressed: () async{
-                  await note.delete();
+                onPressed: () {
+                  showDialog(
+                      context: context,
+                      builder: (context){
+                        return AlertDialog(
+                          title: const Text("Sure you want to delete this note ?"),
+                          actions: [
+                            TextButton(
+                                onPressed: (){
+                                  Navigator.pop(context);
+                                },
+                                child: const Text('Cancel')
+                            ),
+                            TextButton(
+                                onPressed: (){
+                                  note.delete();
+                                  BlocProvider.of<ShowNotesCubit>(context).showAllNotes();
+                                  Navigator.pop(context);
+                                },
+                                child: const Text('Delete')
+                            )
+                          ],
+                        );
+                      }
+                  );
                 },
                 icon: const Icon(
                   FontAwesomeIcons.trash,
